worker: rename to worker_threads
