+++ benchmark/fixtures/echo.worker.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { parentPort } = require('worker');
+const { parentPort } = require('worker_threads');
 
 parentPort.on('message', (msg) => {
   parentPort.postMessage(msg);
