+++ lib/internal/worker.js
@@ -387,7 +387,7 @@ function setupChild(evalScript) {
   debug(`[${threadId}] is setting up worker child environment`);
   const port = getEnvMessagePort();
 
-  const publicWorker = require('worker');
+  const publicWorker = require('worker_threads');
 
   port.on('message', (message) => {
     if (message.type === 'loadScript') {
