+++ lib/worker_threads.js
