+++ test/fixtures/worker-script.mjs
@@ -1,3 +1,3 @@
-import worker from 'worker';
+import worker from 'worker_threads';
 
 worker.parentPort.postMessage('Hello, world!');
