+++ test/parallel/test-message-channel-sharedarraybuffer.js
@@ -3,14 +3,14 @@
 
 const common = require('../common');
 const assert = require('assert');
-const { Worker } = require('worker');
+const { Worker } = require('worker_threads');
 
 {
   const sharedArrayBuffer = new SharedArrayBuffer(12);
   const local = Buffer.from(sharedArrayBuffer);
 
   const w = new Worker(`
-    const { parentPort } = require('worker');
+    const { parentPort } = require('worker_threads');
     parentPort.on('message', ({ sharedArrayBuffer }) => {
       const local = Buffer.from(sharedArrayBuffer);
       local.write('world!', 6);
