+++ test/parallel/test-message-port-message-port-transferring.js
@@ -3,7 +3,7 @@
 const common = require('../common');
 const assert = require('assert');
 
-const { MessageChannel } = require('worker');
+const { MessageChannel } = require('worker_threads');
 
 {
   const { port1: basePort1, port2: basePort2 } = new MessageChannel();
