+++ test/parallel/test-worker-dns-terminate.js
@@ -1,12 +1,12 @@
 // Flags: --experimental-worker
 'use strict';
 const common = require('../common');
-const { Worker } = require('worker');
+const { Worker } = require('worker_threads');
 
 const w = new Worker(`
 const dns = require('dns');
 dns.lookup('nonexistent.org', () => {});
-require('worker').parentPort.postMessage('0');
+require('worker_threads').parentPort.postMessage('0');
 `, { eval: true });
 
 w.on('message', common.mustCall(() => {
