+++ test/parallel/test-worker-esmodule.js
@@ -3,7 +3,7 @@
 const common = require('../common');
 const fixtures = require('../common/fixtures');
 const assert = require('assert');
-const { Worker } = require('worker');
+const { Worker } = require('worker_threads');
 
 const w = new Worker(fixtures.path('worker-script.mjs'));
 w.on('message', common.mustCall((message) => {
