+++ test/parallel/test-worker-syntax-error-file.js
@@ -3,7 +3,7 @@
 const common = require('../common');
 const fixtures = require('../common/fixtures');
 const assert = require('assert');
-const { Worker } = require('worker');
+const { Worker } = require('worker_threads');
 
 // Do not use isMainThread so that this test itself can be run inside a Worker.
 if (!process.env.HAS_STARTED_WORKER) {
