+++ test/parallel/test-worker-uncaught-exception.js
@@ -2,7 +2,7 @@
 'use strict';
 const common = require('../common');
 const assert = require('assert');
-const { Worker } = require('worker');
+const { Worker } = require('worker_threads');
 
 // Do not use isMainThread so that this test itself can be run inside a Worker.
 if (!process.env.HAS_STARTED_WORKER) {
